Refactoring #1: creates Target class, refactors personal information, work experience, education, and websites to use targets
Detailed design notes: The problem that the design solves is: there's lots of repeated code throughout, there is basically no enforced structure, and low extensibility (to new sections, modified logic). The code as it was just had a bunch of functions for each section in the application, with each function coupling the section to be filled and the filling logic. These refactors aim to decouple the section to fill from the filling logic, and fix the aforementioned problems by creating a defined structure that improves flexibility and reduces code duplication. The design is: every time the script runs, it searches for all its Targets (the lowest level unit of what to fill - in order to prevent large 'umbrella'-like targets from preventing nested elements inside it from being unfilled - with a root-level (not nested) selector) and attempts to fill them. This way, the target itself and the current view/page are decoupled. A Target is composed of a root element to locate, value to populate, and the filling function handle - enabling custom logic for a target.